Extracted the results into a different component
